+++ services/multi-vault/keystore/account/exported.go
@@ -1,0 +1,3 @@
+package account
+
+const component = "account"