+++ services/multi-vault/keystore/wallet/exported.go
@@ -1,3 +1,0 @@
-package wallet
-
-const component = "wallet"